feat: add token symbol utility function
