+++ app/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './errorUtil'
 export * from './generalUtil'
+export * from './constantUtil'
